requirements.txt
langchain
langchain-openai
langchain-anthropic
openai
anthropic
python-dotenv
fastapi
uvicorn
playwright
pandas
pydantic